# [core] Performance: remove memory allocations