--- a/packages/react/src/popover/arrow/PopoverArrow.tsx
+++ b/packages/react/src/popover/arrow/PopoverArrow.tsx
@@ -32,14 +32,14 @@ export const PopoverArrow = React.forwardRef(function PopoverArrow(
     [open, side, align, arrowUncentered],
   );
 
-  const renderElement = useRenderElement('div', componentProps, {
+  const element = useRenderElement('div', componentProps, {
     state,
     ref: [forwardedRef, arrowRef],
     props: [{ style: arrowStyles, 'aria-hidden': true }, elementProps],
     customStyleHookMapping: popupStateMapping,
   });
 
-  return renderElement();
+  return element;
 });
 
 export namespace PopoverArrow {