--- a/packages/react/src/popover/positioner/PopoverPositioner.tsx
+++ b/packages/react/src/popover/positioner/PopoverPositioner.tsx
@@ -77,7 +77,7 @@ export const PopoverPositioner = React.forwardRef(function PopoverPositioner(
     [open, positioner.side, positioner.align, positioner.anchorHidden],
   );
 
-  const renderElement = useRenderElement('div', componentProps, {
+  const element = useRenderElement('div', componentProps, {
     state,
     props: [positioner.props, elementProps],
     ref: [forwardedRef, setPositionerElement],
@@ -89,7 +89,7 @@ export const PopoverPositioner = React.forwardRef(function PopoverPositioner(
       {mounted && modal === true && openReason !== 'hover' && openMethod !== 'touch' && (
         <InternalBackdrop inert={inertValue(!open)} />
       )}
-      {renderElement()}
+      {element}
     </PopoverPositionerContext.Provider>
   );
 });