--- a/packages/react/src/utils/useEventCallback.ts
+++ b/packages/react/src/utils/useEventCallback.ts
@@ -11,6 +11,7 @@ const ASSERT_NOT_CALLED = () => {
 };
 
 type Callback =
+  | (() => any)
   | ((a: any) => any)
   | ((a: any, b: any) => any)
   | ((a: any, b: any, c: any) => any)