Refactor a few low-level utilities to avoid hitting the Garbage Collector so much.

Methodology to find optimization targets: check which functions are allocating large amount of objects via the memory profiler:
<img src="https://github.com/user-attachments/assets/d993d33d-69d8-4c4d-9197-30325d5f585c" width="500" />

## Test case

Run the mui-next docs, start the profiler, and navigate to the `/select` page. Code is warmed up before the profile is ran by mounting/unmounting the page a few times. No CPU slowdown is set as it adds noise from one run to another.

## Results

> [!WARNING]
> Benchmarking the GC is imprecise and non-deterministic.

The various individual profiles that I ran result in a 15-35% decrease in total render time.

| Before | After |
| --- | --- |
| ![Screenshot From 2025-05-12 19-23-12](https://github.com/user-attachments/assets/400f0068-c1b9-42dc-bbde-9070150d3bcb) | ![Screenshot From 2025-05-12 19-20-52](https://github.com/user-attachments/assets/94727b27-3638-4953-8569-c150a797aa0e) |