@romgrk 2025-05-11
> It seems better if these changes were made in floating-ui

Yes, but I first need to prove that the changes provide measurable gains, it's easier to make the changes here and test with our internal repositories (joy-ui mainly).

I also don't love having the code live in an external repository. It might be nice to create instead a small hooks package in the mui org, and then send a PR to floating-ui to use it? It's not the first time we mention a hooks package, and it could also benefit material-ui. The versions implemented here are faster than some of the hooks I wrote/tweaked in material-ui.