@flaviendelangle 2025-05-11
Floating UI is maintained mostly by @atomiks