@atomiks 2025-05-15
Fair enough, 6 plus is pretty excessive indeed.